Server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
pymysql
torch
torchvision
pillow
cryptography
python-multipart